refactor: 네이밍 변경 + 간단한 수정
