+++ src/main/java/kr/codesquad/view/OutputView.java
@@ -1,13 +1,12 @@
-package kr.codesquad;
+package kr.codesquad.view;
 
 import java.util.List;
 
-public class Printer {
+public class OutputView {
     public void printResult(List<List<String>> ladderBoard, List<String> namesList) {
         printNames(namesList);
         printLadder(ladderBoard);
     }
-
     private void printNames(List<String> namesList) {
         StringBuilder makeNamesOneString = new StringBuilder();
         makeNamesOneString.append(" ");
@@ -17,13 +16,11 @@ public class Printer {
         }
         System.out.println(makeNamesOneString);
     }
-
     private void printLadder(List<List<String>> ladderBoard) {
         for (int ladderFloorNum = 0; ladderFloorNum < ladderBoard.size(); ladderFloorNum++) {
             printLadderFloorElements(ladderBoard, ladderFloorNum);
         }
     }
-
     private void printLadderFloorElements(List<List<String>> ladderBoard, int ladderFloorNum) {
         System.out.print("    ");
         for (String ladderFloorElements : ladderBoard.get(ladderFloorNum)) {
